First attempt at a chunking store.
Model is to store chunked sequence in a dedicated table. However the implementation is flawed as the output from the store method would be multiple objects, which consumers currently assume will be one (an instance of RawSeq). There needs to be an inversion of control during storage to allow the seq store layer to add to the sqlalchmey session. This might need to happen anyway to support third party/alternative stores